Fixa bilprefab

+++ AbandonGame/Assets/Scripts/CarController.cs
@@ -117,7 +117,7 @@ public class CarController : MonoBehaviour {
         wheelRL.transform.Rotate(wheelColRL.rpm * 60 * Time.deltaTime, 0, 0);
         wheelRR.transform.Rotate(wheelColRR.rpm * 60 * Time.deltaTime, 0, 0);
 
-        engAudio.pitch = rigidB.velocity.magnitude / 32 + (float)0.6;
+        engAudio.pitch = rigidB.velocity.magnitude / 32 + 0.6f;
     }
 
     void FixedUpdate () {
